+++ test/req/integration_test.exs
@@ -25,7 +25,6 @@ defmodule Req.IntegrationTest do
   doctest Req.Steps,
     only: [
       auth: 1,
-      encode_headers: 1,
       put_default_user_agent: 1,
       compressed: 1,
       put_base_url: 1,